Move fold SQL logic into SqlDataFrame
