+++ vegafusion-rt-datafusion/src/sql/dataframe/mod.rs
@@ -233,6 +233,128 @@ impl SqlDataFrame {
             .with_context(|| "unsupported limit query".to_string())
     }
 
+    pub async fn fold(
+        &self,
+        fields: &[String],
+        value_col: &str,
+        key_col: &str,
+        order_field: Option<&str>,
+    ) -> Result<Arc<Self>> {
+        let dialect = self.dialect();
+
+        // Build selection that includes all input fields that aren't shadowed by key/value cols
+        let input_selection = self
+            .schema()
+            .fields()
+            .iter()
+            .filter_map(|f| {
+                if f.name() == &key_col || f.name() == &value_col {
+                    None
+                } else {
+                    Some(flat_col(f.name()))
+                }
+            })
+            .collect::<Vec<_>>();
+
+        // Build query per field
+        let subqueries = fields
+            .iter()
+            .enumerate()
+            .map(|(i, field)| {
+                // Clone input selection and add key/val cols to it
+                let mut subquery_selection = input_selection.clone();
+                subquery_selection.push(lit(field).alias(key_col.clone()));
+                if self.schema().column_with_name(field).is_some() {
+                    // Field exists as a column in the parent table
+                    subquery_selection.push(flat_col(field).alias(value_col.clone()));
+                } else {
+                    // Field does not exist in parent table, fill in NULL instead
+                    subquery_selection.push(lit(ScalarValue::Null).alias(value_col.clone()));
+                }
+
+                if let Some(order_field) = order_field {
+                    let field_order_col = format!("{order_field}_field");
+                    subquery_selection.push(lit(i as u32).alias(field_order_col.clone()));
+                }
+
+                // Create selection CSV for subquery
+                let selection_strs = subquery_selection
+                    .iter()
+                    .map(|sel| Ok(sel.to_sql_select()?.sql(dialect)?))
+                    .collect::<Result<Vec<_>>>()?;
+                let selection_csv = selection_strs.join(", ");
+
+                Ok(format!(
+                    "(SELECT {selection_csv} from {parent})",
+                    selection_csv = selection_csv,
+                    parent = self.parent_name()
+                ))
+            })
+            .collect::<Result<Vec<_>>>()?;
+
+        let union_subquery = subqueries.join(" UNION ALL ");
+        let union_subquery_name = "_union";
+
+        let mut selections = input_selection.clone();
+        selections.push(flat_col(&key_col));
+        selections.push(flat_col(&value_col));
+        if let Some(order_field) = order_field {
+            let field_order_col = format!("{order_field}_field");
+            selections.push(flat_col(&field_order_col));
+        }
+
+        let selection_strs = selections
+            .iter()
+            .map(|sel| Ok(sel.to_sql_select()?.sql(dialect)?))
+            .collect::<Result<Vec<_>>>()?;
+        let selection_csv = selection_strs.join(", ");
+
+        let sql =
+            format!("SELECT {selection_csv} FROM ({union_subquery}) as {union_subquery_name}");
+        let dataframe = self.chain_query_str(&sql).await?;
+
+        if let Some(order_field) = order_field {
+            // Add new ordering column, ordering by:
+            // 1. input row ordering
+            // 2. field index
+            let field_order_col = format!("{order_field}_field");
+            let order_col = Expr::WindowFunction(expr::WindowFunction {
+                fun: window_function::WindowFunction::BuiltInWindowFunction(
+                    BuiltInWindowFunction::RowNumber,
+                ),
+                args: vec![],
+                partition_by: vec![],
+                order_by: vec![
+                    Expr::Sort(expr::Sort {
+                        expr: Box::new(flat_col(order_field)),
+                        asc: true,
+                        nulls_first: false,
+                    }),
+                    Expr::Sort(expr::Sort {
+                        expr: Box::new(flat_col(&field_order_col)),
+                        asc: true,
+                        nulls_first: false,
+                    }),
+                ],
+                window_frame: WindowFrame {
+                    units: WindowFrameUnits::Rows,
+                    start_bound: WindowFrameBound::Preceding(ScalarValue::UInt64(None)),
+                    end_bound: WindowFrameBound::CurrentRow,
+                },
+            })
+            .alias(order_field);
+
+            // Build output selections
+            let mut selections = input_selection.clone();
+            selections.push(flat_col(&key_col));
+            selections.push(flat_col(&value_col));
+            selections[0] = order_col;
+            dataframe.select(selections).await
+        } else {
+            Ok(dataframe)
+        }
+    }
+
     pub async fn impute(
         &self,
         field: &str,
